Moving Operations to own files
